requirements.txt
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
dask==1.2.2
distributed==1.28.1
HeapDict==1.0.1
jmetalpy==1.5.4
joblib==0.14.0
kiwisolver==1.1.0
locket==0.2.0
matplotlib==3.1.2
msgpack==0.6.2
numpy==1.17.4
pandas==0.25.3
partd==1.1.0
patsy==0.5.1
plotly==4.3.0
psutil==5.6.5
py4j==0.10.7
pyparsing==2.4.5
pyspark==2.4.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
retrying==1.3.3
scikit-learn==0.21.3
scipy==1.3.0
six==1.13.0
sklearn==0.0
sortedcontainers==2.1.0
statsmodels==0.10.2
tblib==1.5.0
toolz==0.10.0
tornado==6.0.3
tqdm==4.39.0
typing==3.7.4.1
zict==1.0.0
